+++ tests/test_converter.py
@@ -16,6 +16,7 @@ from typing import (
     Union,
 )
 
+import attrs
 import pytest
 from attrs import Factory, define, field, fields, has, make_class
 from hypothesis import assume, given
@@ -336,6 +337,64 @@ def test_omit_default_roundtrip(cl_and_vals):
     inst = C(0)
     unstructured = converter.unstructure(inst)
     assert unstructured == {"a": 0}
+    assert inst == converter.structure(unstructured, C)
+
+
+@given(simple_typed_classes(defaults="always", allow_nan=False))
+def test_omit_default_with_attrs_converter_roundtrip(cl_and_vals):
+    """
+    Omit default with an attrs converter works.
+    """
+    converter = Converter(omit_if_default=True)
+    cl, vals, kwargs = cl_and_vals
+
+    @define
+    class C:
+        a1: int = field(default="1", converter=int)
+        a2: int = field(default="1", converter=attrs.Converter(int))
+        a3: int = field(factory=lambda: "1", converter=int)
+        a4: int = field(factory=lambda: "1", converter=attrs.Converter(int))
+        a5: int = field(
+            factory=lambda: "2",
+            converter=attrs.Converter(
+                lambda obj, self: int(obj) + self.a4, takes_self=True
+            ),
+        )
+        a6: int = field(
+            factory=lambda: "2",
+            converter=attrs.Converter(
+                lambda obj, field: int(obj) + int(field.default.factory()) - 2,
+                takes_field=True,
+            ),
+        )
+        a7: int = field(
+            factory=lambda: "3",
+            converter=attrs.Converter(
+                lambda obj, self, field: (
+                    int(obj) + self.a6 + int(field.default.factory()) - 3
+                ),
+                takes_self=True,
+                takes_field=True,
+            ),
+        )
+        c: cl = Factory(lambda: cl(*vals, **kwargs))
+
+    inst = C()
+    unstructured = converter.unstructure(inst)
+    assert unstructured == {}
+    assert inst == converter.structure(unstructured, C)
+
+    inst = C(0, 0, 0, 0, 0, 0, 0)
+    unstructured = converter.unstructure(inst)
+    assert unstructured == {
+        "a1": 0,
+        "a2": 0,
+        "a3": 0,
+        "a4": 0,
+        "a5": 0,
+        "a6": 0,
+        "a7": 0,
+    }
     assert inst == converter.structure(unstructured, C)