+++ HISTORY.md
@@ -17,6 +17,8 @@ Our backwards-compatibility policy can be found [here](https://github.com/python
   ([#688](https://github.com/python-attrs/cattrs/pull/688))
 - Python 3.9 is no longer supported, as it is end-of-life. Use previous versions on this Python version.
   ([#698](https://github.com/python-attrs/cattrs/pull/698))
+- Apply the attrs converter to the default value before checking if it is equal to the attribute's value, when `omit_if_default` is true and an attrs converter is specified.
+  ([#696](https://github.com/python-attrs/cattrs/pull/696))
 
 ## 25.3.0 (2025-10-07)